+++ examples/custom_exception_ruby/ext/ahriman/src/lib.rs
@@ -1,15 +1,15 @@
 use magnus::{
     class, define_module, exception::ExceptionClass, function, gc::register_mark_object, memoize,
-    prelude::*, Error, RClass, RModule,
+    prelude::*, Error, RModule,
 };
 
 fn rubric_error() -> ExceptionClass {
     *memoize!(ExceptionClass: {
-        let ex: RClass = class::object().const_get::<_, RModule>("Ahriman").unwrap().const_get("RubricError").unwrap();
+        let ex = class::object().const_get::<_, RModule>("Ahriman").unwrap().const_get("RubricError").unwrap();
         // ensure `ex` is never garbage collected (e.g. if constant is
         // redefined) and also not moved under compacting GC.
         register_mark_object(ex);
-        ExceptionClass::from_value(*ex).unwrap()
+        ex
     })
 }